+++ test/td_ie/ingests/workflow_test.exs
@@ -5,7 +5,7 @@ defmodule TdIe.Ingests.WorkflowTest do
   import TdIe.TestOperators
 
   alias TdCache.Redix.Stream
-  alias TdCore.Search.IndexWorkerMock
+  alias TdCore.Search.IndexWorker
   alias TdIe.Ingests
   alias TdIe.Ingests.IngestVersion
   alias TdIe.Ingests.Workflow
@@ -31,7 +31,7 @@ defmodule TdIe.Ingests.WorkflowTest do
     setup :put_user
 
     test "with valid data creates a ingest", %{user_id: user_id} do
-      IndexWorkerMock.clear()
+      IndexWorker.clear()
       domain_id = 1
 
       ingest_attrs = %{
@@ -65,12 +65,12 @@ defmodule TdIe.Ingests.WorkflowTest do
       assert ingest_version.ingest.type == ingest_attrs.type
       assert ingest_version.ingest.domain_id == ingest_attrs.domain_id
       assert ingest_version.ingest.last_change_by == ingest_attrs.last_change_by
-      assert [{:reindex, :ingests, [_]} | _] = IndexWorkerMock.calls()
-      IndexWorkerMock.clear()
+      assert [{:reindex, :ingests, [_]} | _] = IndexWorker.calls()
+      IndexWorker.clear()
     end
 
     test "publishes an audit event including domain_id", %{user_id: user_id} do
-      IndexWorkerMock.clear()
+      IndexWorker.clear()
       domain_id = 1
 
       ingest_attrs = %{
@@ -96,8 +96,8 @@ defmodule TdIe.Ingests.WorkflowTest do
       assert {:ok, [%{id: ^event_id} = event]} = Stream.read(:redix, @stream, transform: true)
       assert %{payload: payload} = event
       assert %{"ingest" => %{"domain_id" => ^domain_id}} = Jason.decode!(payload)
-      assert [{:reindex, :ingests, [_]}] = IndexWorkerMock.calls()
-      IndexWorkerMock.clear()
+      assert [{:reindex, :ingests, [_]}] = IndexWorker.calls()
+      IndexWorker.clear()
     end
 
     test "with invalid data returns error changeset" do
@@ -118,7 +118,7 @@ defmodule TdIe.Ingests.WorkflowTest do
     end
 
     test "with content", %{user_id: user_id} do
-      IndexWorkerMock.clear()
+      IndexWorker.clear()
       domain_id = 1
 
       content_schema = [
@@ -151,12 +151,12 @@ defmodule TdIe.Ingests.WorkflowTest do
       assert {:ok, %{ingest_version: ingest_version}} = Workflow.create_ingest(creation_attrs)
 
       assert %{content: ^content} = ingest_version
-      assert [{:reindex, :ingests, [_]}] = IndexWorkerMock.calls()
-      IndexWorkerMock.clear()
+      assert [{:reindex, :ingests, [_]}] = IndexWorker.calls()
+      IndexWorker.clear()
     end
 
     test "with invalid content: required", %{user_id: user_id} do
-      IndexWorkerMock.clear()
+      IndexWorker.clear()
       domain_id = 1
 
       content_schema = [
@@ -197,12 +197,12 @@ defmodule TdIe.Ingests.WorkflowTest do
       assert ingest_version.ingest.type == ingest_attrs.type
       assert ingest_version.ingest.domain_id == ingest_attrs.domain_id
       assert ingest_version.ingest.last_change_by == ingest_attrs.last_change_by
-      assert [{:reindex, :ingests, [_]}] = IndexWorkerMock.calls()
-      IndexWorkerMock.clear()
+      assert [{:reindex, :ingests, [_]}] = IndexWorker.calls()
+      IndexWorker.clear()
     end
 
     test "with content: default values", %{user_id: user_id} do
-      IndexWorkerMock.clear()
+      IndexWorker.clear()
       domain_id = 1
 
       content_schema = [
@@ -235,12 +235,12 @@ defmodule TdIe.Ingests.WorkflowTest do
 
       assert ingest_version.content["Field1"] == "Hello"
       assert ingest_version.content["Field2"] == "World"
-      assert [{:reindex, :ingests, [_]} | _] = IndexWorkerMock.calls()
-      IndexWorkerMock.clear()
+      assert [{:reindex, :ingests, [_]} | _] = IndexWorker.calls()
+      IndexWorker.clear()
     end
 
     test "with invalid content: invalid variable list", %{user_id: user_id} do
-      IndexWorkerMock.clear()
+      IndexWorker.clear()
       domain_id = 1
 
       content_schema = [
@@ -280,8 +280,8 @@ defmodule TdIe.Ingests.WorkflowTest do
       assert ingest_version.ingest.type == ingest_attrs.type
       assert ingest_version.ingest.domain_id == ingest_attrs.domain_id
       assert ingest_version.ingest.last_change_by == ingest_attrs.last_change_by
-      assert [{:reindex, :ingests, [_]}] = IndexWorkerMock.calls()
-      IndexWorkerMock.clear()
+      assert [{:reindex, :ingests, [_]}] = IndexWorker.calls()
+      IndexWorker.clear()
     end
 
     test "with no content", %{user_id: user_id} do
@@ -377,7 +377,7 @@ defmodule TdIe.Ingests.WorkflowTest do
     setup :put_user
 
     test "with valid data updates the ingest_version", %{user_id: user_id} do
-      IndexWorkerMock.clear()
+      IndexWorker.clear()
       ingest_version = insert(:ingest_version)
 
       ingest_attrs = %{
@@ -410,12 +410,12 @@ defmodule TdIe.Ingests.WorkflowTest do
 
       assert object.ingest.id == ingest_version.ingest.id
       assert object.ingest.last_change_by == 1000
-      assert [{:reindex, :ingests, [_]}] = IndexWorkerMock.calls()
-      IndexWorkerMock.clear()
+      assert [{:reindex, :ingests, [_]}] = IndexWorker.calls()
+      IndexWorker.clear()
     end
 
     test "with valid content data updates the ingest", %{user_id: user_id} do
-      IndexWorkerMock.clear()
+      IndexWorker.clear()
 
       content_schema = [
         %{"name" => "Field1", "type" => "string", "required" => true},
@@ -455,8 +455,8 @@ defmodule TdIe.Ingests.WorkflowTest do
       assert %IngestVersion{} = ingest_version
       assert ingest_version.content["Field1"] == "New first field"
       assert ingest_version.content["Field2"] == "Second field"
-      assert [{:reindex, :ingests, [_]}] = IndexWorkerMock.calls()
-      IndexWorkerMock.clear()
+      assert [{:reindex, :ingests, [_]}] = IndexWorker.calls()
+      IndexWorker.clear()
     end
 
     test "with invalid data returns error changeset" do
@@ -523,12 +523,12 @@ defmodule TdIe.Ingests.WorkflowTest do
     test "creates a new version and sets current to false on previous version", %{
       claims: claims
     } do
-      IndexWorkerMock.clear()
+      IndexWorker.clear()
       ingest_version = insert(:ingest_version, status: "published")
       assert {:ok, res} = Workflow.new_ingest_version(ingest_version, claims)
       assert %{current: %{current: true}, previous: %{current: false}} = res
-      assert [{:reindex, :ingests, [_]}] = IndexWorkerMock.calls()
-      IndexWorkerMock.clear()
+      assert [{:reindex, :ingests, [_]}] = IndexWorker.calls()
+      IndexWorker.clear()
     end
 
     test "creates a new version and copies the identifier from the previous version one", %{
@@ -536,7 +536,7 @@ defmodule TdIe.Ingests.WorkflowTest do
       template_with_identifier: template_with_identifier,
       identifier_name: identifier_name
     } do
-      IndexWorkerMock.clear()
+      IndexWorker.clear()
       existing_identifier = "00000000-0000-0000-0000-000000000000"
       ingest = build(:ingest, %{type: template_with_identifier.name})
 
@@ -558,17 +558,17 @@ defmodule TdIe.Ingests.WorkflowTest do
                }
              } = res
 
-      assert [{:reindex, :ingests, [_]}] = IndexWorkerMock.calls()
-      IndexWorkerMock.clear()
+      assert [{:reindex, :ingests, [_]}] = IndexWorker.calls()
+      IndexWorker.clear()
     end
 
     test "publishes an event to the audit stream", %{claims: claims} do
-      IndexWorkerMock.clear()
+      IndexWorker.clear()
       ingest_version = insert(:ingest_version, status: "published")
       assert {:ok, %{audit: event_id}} = Workflow.new_ingest_version(ingest_version, claims)
       assert {:ok, [%{id: ^event_id}]} = Stream.read(:redix, @stream, transform: true)
-      assert [{:reindex, :ingests, [_]}] = IndexWorkerMock.calls()
-      IndexWorkerMock.clear()
+      assert [{:reindex, :ingests, [_]}] = IndexWorker.calls()
+      IndexWorker.clear()
     end
   end
 
@@ -576,7 +576,7 @@ defmodule TdIe.Ingests.WorkflowTest do
     setup :claims
 
     test "changes the status and audit fields" do
-      IndexWorkerMock.clear()
+      IndexWorker.clear()
       %{last_change_at: ts} = ingest_version = insert(:ingest_version, status: "draft")
 
       %{user_id: user_id} = claims = build(:claims, user_id: 987)
@@ -588,12 +588,12 @@ defmodule TdIe.Ingests.WorkflowTest do
                published
 
       assert DateTime.diff(last_change_at, ts, :microsecond) > 0
-      assert [{:reindex, :ingests, [_]}] = IndexWorkerMock.calls()
-      IndexWorkerMock.clear()
+      assert [{:reindex, :ingests, [_]}] = IndexWorker.calls()
+      IndexWorker.clear()
     end
 
     test "publishes an event to ingests:events", %{claims: claims} do
-      IndexWorkerMock.clear()
+      IndexWorker.clear()
       %{ingest_id: ingest_id, id: id} = ingest_version = insert(:ingest_version)
       assert {:ok, %{event: event_id}} = Workflow.publish_ingest_version(ingest_version, claims)
 
@@ -601,12 +601,12 @@ defmodule TdIe.Ingests.WorkflowTest do
                Stream.range(:redix, "ingests:events", event_id, event_id, transform: false)
 
       assert event_data ||| ["event", "publish", "id", "#{ingest_id}", "version_id", "#{id}"]
-      assert [{:reindex, :ingests, [_]}] = IndexWorkerMock.calls()
-      IndexWorkerMock.clear()
+      assert [{:reindex, :ingests, [_]}] = IndexWorker.calls()
+      IndexWorker.clear()
     end
 
     test "publishes an event including domain_ids to the audit stream", %{claims: claims} do
-      IndexWorkerMock.clear()
+      IndexWorker.clear()
       ingest_version = insert(:ingest_version, status: "draft")
 
       assert {:ok, %{audit: event_id}} = Workflow.publish_ingest_version(ingest_version, claims)
@@ -614,8 +614,8 @@ defmodule TdIe.Ingests.WorkflowTest do
       assert {:ok, [%{id: ^event_id} = event]} = Stream.read(:redix, @stream, transform: true)
       assert %{payload: payload} = event
       assert %{"domain_ids" => _domain_ids} = Jason.decode!(payload)
-      assert [{:reindex, :ingests, [_]} | _] = IndexWorkerMock.calls()
-      IndexWorkerMock.clear()
+      assert [{:reindex, :ingests, [_]} | _] = IndexWorker.calls()
+      IndexWorker.clear()
     end
   end
 
@@ -623,7 +623,7 @@ defmodule TdIe.Ingests.WorkflowTest do
     setup :claims
 
     test "rejects ingest", %{claims: claims} do
-      IndexWorkerMock.clear()
+      IndexWorker.clear()
       reason = "Because I want to"
       ingest_version = insert(:ingest_version, status: "pending_approval")
 
@@ -631,12 +631,12 @@ defmodule TdIe.Ingests.WorkflowTest do
                Workflow.reject_ingest_version(ingest_version, reason, claims)
 
       assert %{status: "rejected", reject_reason: ^reason} = ingest_version
-      assert [{:reindex, :ingests, [_]}] = IndexWorkerMock.calls()
-      IndexWorkerMock.clear()
+      assert [{:reindex, :ingests, [_]}] = IndexWorker.calls()
+      IndexWorker.clear()
     end
 
     test "publishes an event to the audit stream", %{claims: claims} do
-      IndexWorkerMock.clear()
+      IndexWorker.clear()
       reason = "Because I want to"
       ingest_version = insert(:ingest_version, status: "pending_approval")
 
@@ -644,8 +644,8 @@ defmodule TdIe.Ingests.WorkflowTest do
                Workflow.reject_ingest_version(ingest_version, reason, claims)
 
       assert {:ok, [%{id: ^event_id}]} = Stream.read(:redix, @stream, transform: true)
-      assert [{:reindex, :ingests, [_]}] = IndexWorkerMock.calls()
-      IndexWorkerMock.clear()
+      assert [{:reindex, :ingests, [_]}] = IndexWorker.calls()
+      IndexWorker.clear()
     end
   end
 
@@ -653,19 +653,19 @@ defmodule TdIe.Ingests.WorkflowTest do
     setup :claims
 
     test "updates the ingest", %{claims: %{user_id: user_id} = claims} do
-      IndexWorkerMock.clear()
+      IndexWorker.clear()
       ingest_version = insert(:ingest_version)
 
       assert {:ok, %{updated: ingest_version}} =
                Workflow.submit_ingest_version(ingest_version, claims)
 
       assert %{status: "pending_approval", last_change_by: ^user_id} = ingest_version
-      assert [{:reindex, :ingests, [_]}] = IndexWorkerMock.calls()
-      IndexWorkerMock.clear()
+      assert [{:reindex, :ingests, [_]}] = IndexWorker.calls()
+      IndexWorker.clear()
     end
 
     test "publishes an event including domain_ids to the audit stream", %{claims: claims} do
-      IndexWorkerMock.clear()
+      IndexWorker.clear()
       ingest_version = insert(:ingest_version, status: "draft")
 
       assert {:ok, %{audit: event_id}} = Workflow.submit_ingest_version(ingest_version, claims)
@@ -673,8 +673,8 @@ defmodule TdIe.Ingests.WorkflowTest do
       assert {:ok, [%{id: ^event_id} = event]} = Stream.read(:redix, @stream, transform: true)
       assert %{payload: payload} = event
       assert %{"domain_ids" => _domain_ids} = Jason.decode!(payload)
-      assert [{:reindex, :ingests, [_]}] = IndexWorkerMock.calls()
-      IndexWorkerMock.clear()
+      assert [{:reindex, :ingests, [_]}] = IndexWorker.calls()
+      IndexWorker.clear()
     end
   end