+++ CHANGELOG.md
@@ -1,5 +1,15 @@
 # Changelog
 
+## [6.7.0] 2024-06-13
+
+### Fixed
+
+- [TD-6561]
+  - Standardise aggregations limits
+  - Use keyword list for elastic search configurationç
+- [TD-6402] IndexWorker improvement
+- [TD-6440] Update td-df-lib
+
 ## [6.5.0] 2024-04-30
 
 ### Added